Moved stat_string_coord to UIPopup.cpp
