+++ src/Engine/mm7_data.cpp
@@ -166,23 +166,6 @@ struct IconFrameTable *pIconsFrameTable;
 #include "Engine/Tables/PlayerFrameTable.h"
 struct PlayerFrameTable *pPlayerFrameTable;  // idb
 
-std::array<stat_coord, 26> stat_string_coord =  // 4E2940
-    {{
-        {0x1A, 0x39, 0xDC, 0x12},   {0x1A, 0x4A, 0xDC, 0x12},
-        {0x1A, 0x5B, 0xDC, 0x12},   {0x1A, 0x6C, 0xDC, 0x12},
-        {0x1A, 0x7D, 0xDC, 0x12},   {0x1A, 0x8E, 0xDC, 0x12},
-        {0x1A, 0x9F, 0xDC, 0x12},   {0x1A, 0xC6, 0xDC, 0x12},
-        {0x1A, 0xD7, 0xDC, 0x12},   {0x1A, 0xE8, 0xDC, 0x12},
-        {0x1A, 0x10C, 0xDC, 0x12},  {0x1A, 0x11E, 0xDC, 0x12},
-        {0x111, 0x36, 0xBA, 0x12},  {0x111, 0x47, 0xBA, 0x12},
-        {0x111, 0x58, 0xBA, 0x12},  {0x111, 0x7E, 0xBA, 0x12},
-        {0x111, 0x8F, 0xBA, 0x12},  {0x111, 0xA0, 0xBA, 0x12},
-        {0x111, 0xB1, 0xBA, 0x12},  {0x111, 0xCA, 0xBA, 0x12},
-        {0x111, 0xDD, 0xBA, 0x12},  {0x111, 0xF0, 0xBA, 0x12},
-        {0x111, 0x103, 0xBA, 0x12}, {0x111, 0x116, 0xBA, 0x12},
-        {0x111, 0x129, 0xBA, 0x12}, {0x13E, 0x12, 0x89, 0x12},
-    }};
-
 //-------------------------------------------------------------------------
 // Data declarations
